configured pyproject.toml with hatch configuration. Moved the source code under src/strategy_tester
