+++ .gitignore
@@ -24,3 +24,4 @@ doc/build/*
 
 example/
 .env
+.pytest_cache
