+++ README.md
@@ -1,5 +1,7 @@
 # Strategy tester
 
+[![Hatch project](https://img.shields.io/badge/%F0%9F%A5%9A-Hatch-4051b5.svg)](https://github.com/pypa/hatch)
+
 A tool to simplify backtesing and/or optimization of your trading strategy.
 
 ## ByBit
