+++ src/strategy_tester/backtesting/optimization.py
