+++ src/strategy_tester/backtesting/pipeline/factory.py
