+++ src/strategy_tester/backtesting/report_html.py
