+++ src/strategy_tester/metrics/entries_counts.py
