+++ src/strategy_tester/metrics/profits_losses_bars.py
