+++ src/strategy_tester/metrics/profits_losses_by_bar_opened.py
