+++ src/strategy_tester/metrics/report_pdf.py
