+++ src/strategy_tester/metrics/save.py
