+++ src/strategy_tester/metrics/utils.py
