+++ src/strategy_tester/pipeline/context.py
