+++ src/strategy_tester/pipeline/pipe.py
