+++ src/strategy_tester/tests/metrics/entries_counts_test.py
