+++ src/strategy_tester/tests/metrics/profits_losses_bars_test.py
