+++ src/strategy_tester/tests/metrics/profits_losses_by_bar_opened_test.py
