+++ src/strategy_tester/tests/metrics/profits_losses_mean_test.py
