+++ src/strategy_tester/tests/metrics/utils_test.py
