+++ src/strategy_tester/tests/pipeline_test.py
@@ -3,8 +3,7 @@ from dotenv import dotenv_values
 from backtesting import Strategy
 import pytest
 
-from src.broker_params import BrokerParamsBuilder
-
+from ..broker_params import BrokerParamsBuilder
 from ..backtesting.pipeline.steps import get_add_asset_name, get_add_strategy_name, get_add_telegram_bot, get_add_broker_params
 from ..pipeline.context import Context
 from ..pipeline.pipe import pipe
