+++ src/strategy_tester/utils/files.py
